+++ On Your Bike - Chapter 9/src/com/androiddevbook/onyourbike/chapter9/model/Trip.java
@@ -68,7 +68,7 @@ public class Trip {
         database.execSQL(dropTable);
     }
 
-    public void setid(int id) {
+    public void setid(long id) {
         _id = id;
     }
